Fixed last stuff in fladuino make files. Now everything is set up to default values in common.mk and they can be overwritten in the individual makefiles. DEVICE is changed to PLATFORM and this is the main identifyer for the default settings of each board.
Also made a symlink in the fladuino folder so the examples folder in there points to the flask folder in the example-code folder so all examples are
located in a common place, which makes the make-all script easier to deal with. Though this should be changed to a makefile that can handle cleanup of alle
the examples code as well.

Updated the common.mk makefile for all fladuino exmaples to reflect the new auto dependency generation of arduino code. This is a new changed in arduino svn
repository.